requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.13.0
keras==2.13.0
numpy==1.24.0
Pillow==10.2.0
nltk==3.8.1
scikit-learn==1.3.0
gdown==5.1.0
h5py==3.9.0